[CL-1253] Updates from code review
