+++ front/app/components/Form/Components/Controls/index.tsx
@@ -1,0 +1,89 @@
+// renderers
+import CLCategoryLayout, {
+  clCategoryTester,
+} from 'components/Form/Components/Layouts/CLCategoryLayout';
+import OrderedLayout, {
+  orderedLayoutTester,
+} from 'components/Form/Components/Layouts/OrderedLayout';
+import InputControl, {
+  inputControlTester,
+} from 'components/Form/Components/Controls/InputControl';
+import TextAreaControl, {
+  textAreaControlTester,
+} from 'components/Form/Components/Controls/TextAreaControl';
+import WYSIWYGControl, {
+  WYSIWYGControlTester,
+} from 'components/Form/Components/Controls/WYSIWYGControl';
+import TopicsControl, {
+  topicsControlTester,
+} from 'components/Form/Components/Controls/TopicsControl';
+import ImageControl, {
+  imageControlTester,
+} from 'components/Form/Components/Controls/ImageControl';
+import AttachmentsControl, {
+  attachmentsControlTester,
+} from 'components/Form/Components/Controls/AttachmentsControl';
+import DescriptionControl, {
+  descriptionControlTester,
+} from 'components/Form/Components/Controls/DescriptionControl';
+import TitleControl, {
+  titleControlTester,
+} from 'components/Form/Components/Controls/TitleControl';
+import SingleSelectControl, {
+  singleSelectControlTester,
+} from 'components/Form/Components/Controls/SingleSelectControl';
+import MultiSelectControl, {
+  multiSelectControlTester,
+} from 'components/Form/Components/Controls/MultiSelectControl';
+import UserPickerControl, {
+  userPickerControlTester,
+} from 'components/Form/Components/Controls/UserPickerControl';
+import CheckboxControl, {
+  checkboxControlTester,
+} from 'components/Form/Components/Controls/CheckboxControl';
+import LocationControl, {
+  locationControlTester,
+} from 'components/Form/Components/Controls/LocationControl';
+import DateControl, {
+  dateControlTester,
+} from 'components/Form/Components/Controls/DateControl';
+import MultilocInputLayout, {
+  multilocInputTester,
+} from 'components/Form/Components/Controls/MultilocInputLayout';
+
+export {
+  CLCategoryLayout,
+  clCategoryTester,
+  OrderedLayout,
+  orderedLayoutTester,
+  InputControl,
+  inputControlTester,
+  TextAreaControl,
+  textAreaControlTester,
+  WYSIWYGControl,
+  WYSIWYGControlTester,
+  TopicsControl,
+  topicsControlTester,
+  ImageControl,
+  imageControlTester,
+  AttachmentsControl,
+  attachmentsControlTester,
+  DescriptionControl,
+  descriptionControlTester,
+  TitleControl,
+  titleControlTester,
+  SingleSelectControl,
+  singleSelectControlTester,
+  MultiSelectControl,
+  multiSelectControlTester,
+  UserPickerControl,
+  userPickerControlTester,
+  CheckboxControl,
+  checkboxControlTester,
+  LocationControl,
+  locationControlTester,
+  DateControl,
+  dateControlTester,
+  MultilocInputLayout,
+  multilocInputTester,
+};
