+++ front/app/modules/commercial/native_surveys/components/NativeSurveyLayout/index.tsx
